style: visual edit
Ended up changing how the homepage image looked. Added some padding
to make it look more focused. Had to make some changes to the width
of main because it is out of normal flow due to absolute
positioning. If I just do 100% it will go out of the screen, so I
ended up doing 100% - 25vw since that is the size of the header
element.