fix: responsive layout and container overflow
The changes to the initial homepage image broke the responsiveness
of the layout. Editing the media queries by touching up the
positioning and changing the object fit rule for image helped.

There was also an issue where the header bar would move even
though it was fixed in position. To fix this I just added a rule
for overflow to main.