feat: responsive member info text
Change the value of the font-size from rem units to vh units.
This makes the text scale responsively with the viewport size.
Usually don't want to use vh or vw for text size since it
makes it harder to access for visually impared people, but in
my case I need it to maintain consistency in the 2x2 grid display
for the text being displayed.